fix: Prevent pixels not filling when no color is selected
Initializing the colorChoice variable to the default value allows the grid elements to be filled when the user doesn't explicitly pick a colour.  The default colour appears to be black, but I've programmed it to be fetched from the colour picker as opposed to hardcoding in "black" in case there are cases in which the default colour is not black.