fix: Pixel hover effect always being black on new grid instantiation
The bug ocurred when the user selected a non-black colour, then made a new grid.  Because the newly created grid is instantiated with a black hover effect, the colour of this hover effect has now been updated to correctly reflect the current colour selected.  The result is that a newly created grid will have a hover effect colour that matches the currently selected colour, whatever it may be.