feat: Restrict keys available on input fields
Currently only the use of numeric, backspace and delete are available.  This is necessary as WASD were binded to creating rows and columns, which means the validation of the input fields is skipped.  This can result in rows and columns not being properly rendered because of letters in the input fields.  The values can't be sanitised, because if the input is not valid on number inputs (e.g. contains letters), then Firefox doesn't store the invalid value, making it impossible to manipulate that value.