feat: Add a wider range of valid keys on the input field
I'm not completely happy with this solution, as the inputs get validated twice now.  However, as the keybindings for adding rows and columns bypass the form validation, I can't think of a better way right now.  This solution produces the ideal user experience I am looking for, so for the moment, it will do.